Resolved some platforms incompatibilities

+++ src/libq/libq.h
@@ -8,7 +8,7 @@
 namespace libq {
 
 typedef std::complex<float> cmplx;
-typedef uint64_t state_t;
+typedef unsigned long long state_t;
 #define HASH_CACHE_SIZE (1024 * 64)
 
 struct qureg_t {
